feat: reusable pending tx fee component
