+++ components/brave_wallet_ui/components/extension/confirm-transaction-panel/swap.style.tsx
@@ -5,7 +5,6 @@
 
 import styled from 'styled-components'
 
-import { SettingsAdvancedIcon } from 'brave-ui/components/icons'
 import ArrowDown2Icon from '../../../assets/svg-icons/arrow-down-2.svg'
 import { AssetIconFactory, AssetIconProps, WalletButton } from '../../shared/style'
 
@@ -171,50 +170,6 @@ export const NetworkDescriptionText = styled.span`
   font-size: 12px;
   display: flex;
   align-items: center;
-  color: ${p => p.theme.color.text03};
-`
-
-export const NetworkFeeAndSettingsContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  width: calc(100% - 8px);
-`
-
-export const NetworkFeeContainer = styled.div``
-
-export const NetworkFeeTitle = styled.div`
-  font-family: 'Poppins';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 20px;
-  color: ${p => p.theme.color.text03};
-`
-
-export const NetworkFeeValue = styled.div`
-  font-family: 'Poppins';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 18px;
-  display: flex;
-  align-items: center;
-  letter-spacing: 0.01em;
-  color: ${p => p.theme.color.text02};
-  gap: 6px;
-`
-
-export const Settings = styled(WalletButton)`
-  display: flex;
-  align-self: flex-start;
-  cursor: pointer;
-  outline: none;
-  border: none;
-  background: none;
-`
-
-export const SettingsIcon = styled(SettingsAdvancedIcon)`
-  width: 14px;
   color: ${p => p.theme.color.text03};
 `
